Feat: Landing페이지 Card 컴포넌트 생성
